HashCode2014Solution

Documentation for HashCode2014Solution.

HashCode2014Solution.RouteGrid

HashCode2014Solution.RouteGridNode

HashCode2014Solution.ComputeUpperBound

HashCode2014Solution.create_grid

HashCode2014Solution.random_walk_modified

HashCode2014Solution.RouteGrid — Type

RouteGrid

Transform a city into a graph where each junction is a node, and the streets are edges.

Fields

neighbors::Vector{SVector{RouteGridNode}}: for each junction, the list of street-junctions that can be reached from it.

HashCode2014Solution.RouteGridNode — Type

RouteGridNode

Group a junction and a street that leads to it together.

Fields

street::Int: a street in the city

junction::Int: a junction that can be reached from the street

distance::Int: the length of the street

HashCode2014Solution.ComputeUpperBound — Method

ComputeUpperBound(city, duration_limit)

Algorithm to compute upper bound on distance. Compares velocity of each street and constructs 8 separate paths for each car in a "drag race" style journey. This approach relaxes the constraint on connectivity of the streets.

Fields

'city::City': a street in the city

'duration_limit::Int64': the total time limit imposed on cars traversing streets

HashCode2014Solution.create_grid — Method

create_grid(city::City)

Create a RouteGrid from a City.

Arguments

city::City: the city to transform

HashCode2014Solution.random_walk_modified — Method

random_walk_modified(rng, city)
random_walk_modified(city)

Modified version of random_walk() from HashCode2014.