API References

Index

HashCode2014Solution.AdjList

HashCode2014Solution.ModifiedJunction

HashCode2014Solution.ModifiedStreet

HashCode2014Solution.approx_upper_bound

HashCode2014Solution.create_adj_list

HashCode2014Solution.find_best_spreading_duration

HashCode2014Solution.get_direction

HashCode2014Solution.get_modified_junctions

HashCode2014Solution.get_modified_streets

HashCode2014Solution.solve

Documentation

HashCode2014Solution.AdjList — Type

struct AdjList{T}

A struct representing an adjacency list.

Fields

edges::Vector{Vector{T}}: The adjacency list.

total_duration::Int: The total duration of the city.

all_streets::Vector{T}: A vector of all modified streets.

nb_cars::Int: The number of cars.

HashCode2014Solution.ModifiedJunction — Type

struct ModifiedJunction

A struct representing a junction. Similar to Junction from HashCode2014, but with an additional field index representing the index of the junction.

Fields

index::Int: The index of the junction.

latitude::Float64: The latitude of the junction.

longitude::Float64: The longitude of the junction.

HashCode2014Solution.ModifiedStreet — Type

struct ModifiedStreet

A struct representing a street. Similar to Street from HashCode2014 but with additional fields.

Fields

A::ModifiedJunction: The junction at the first endpoint of the street.

B::ModifiedJunction: The junction at the second endpoint of the street.

bidirectional::Bool: Whether the street is bidirectional.

duration::Int: The duration of the street.

distance::Int: The distance of the street.

angle::Int: The angle of the street with respect to the north.

speed::Float16: The speed of the street.

HashCode2014Solution.approx_upper_bound — Method

approx_upper_bound(adj_list::AdjList)

Approximately calculates the upper bound of the distance traveled by nb_cars cars in a given adj_list of streets.

The algorithm works as follows:

Sort the streets by decreasing speed.

For each car, take the fastest street that is not already taken and that does not exceed the total duration of the city.

Add the distance of the selected street to the total distance traveled by the cars.

Repeat steps 2 and 3 until no street can be taken (either because the total duration is exceeded or because all streets are taken).

The algorithm is not optimal since it does not take into account the connectivity of the streets. However, it is fast and gives a good approximation of the upper bound since it is greedily choosing the streets that cover the most distance in the shortest amount of time without violating the duration constraint.

Correctness:

The algorithm terminates since the total duration of the city is finite.

The algorithm is greedy, so it will always choose the street that covers the most distance in the shortest amount of time. Therefore, the returned upper bound is always greater than or equal to the actual possible distance traveled by the cars.

The algorithm does not take into account the connectivity of the streets, so it may not be optimal.

Runtime:

for s streets, c cars, and j junctions:

O(s log(s)) for sorting the streets by decreasing speed

c iterations of the while loop

O(s) for filtering the streets

O(1) for the other operations

Overall runtime: O(s log(s) + c * (s + 1)) = O(s log(s) + c * s)

Arguments

adj_list::AdjList: The adjacency list representation of the street network.

Returns

distance::Int: The approximate upper bound of the total distance traveled by the cars.

HashCode2014Solution.create_adj_list — Method

create_adj_list(city::City)

Create an adjacency list based on the City object.

Arguments

city::City: The City object containing the streets.

Returns

An AdjList object.

HashCode2014Solution.find_best_spreading_duration — Method

find_best_spreading_duration(adj_list::AdjList, resolution::Int)

Find the best spreading duration (that achive the highest total distance) for a given problem graph.

Arguments

adj_list::AdjList: The adjacency list representing the graph.

resolution::Int: The resolution for spreading the duration, i.e. the spreading duration will be incremented by this value at each iteration.

Returns

max_d::Int: The maximum distance achieved.

max_i::Int: The corresponding spreading duration.

Raises

ArgumentError: If the resolution is less than or equal to 0, or greater than the total duration.

HashCode2014Solution.get_direction — Method

get_direction(city::City, A::Int, B::Int)

Compute the angle of the street from junction A to junction B in a given city with respect to the north.

Arguments

city::City: The city object containing the junctions.

A::Int: The index of the starting junction.

B::Int: The index of the destination junction.

Returns

The direction in degrees from junction A to junction B.

HashCode2014Solution.get_modified_junctions — Method

get_modified_junctions(city::City)

Create a list of ModifiedJunction objects based on the City object.

Arguments

city::City: The City object containing the junctions.

Returns

An array of ModifiedJunction objects.

HashCode2014Solution.get_modified_streets — Method

get_modified_streets(city::City)

Create a list of ModifiedStreet objects based on the City object.

Arguments

city::City: The City object containing the streets.

Returns

An array of ModifiedStreet objects.

HashCode2014Solution.solve — Function

solve(adj_list::AdjList{ModifiedStreet})

Solve the HashCode2014 problem using the algorithm described below.

Algorithm

This is a greedy algorithm in which we find the best street to take in each iteration. To do so, we use a few heuristics:

Streets that have been visited fewer times in previous iterations are given higher priority.

Among those, if the car is still in the spreading phase, streets that are directed closer to the car's assigned angle are given higher priority. The car's assigned angle is determined by its index in the list of cars: carangle = carindex * 360/numberofcars.

Among those, streets with higher speed are given higher priority.

A street can only be taken if the car will still be able to reach the end of the street before the end of the simulation. A street cannot be taken if the car is not in the strating end of the street.

Runtime:

The acctual runtime is dependent mostly on the total duration of the simulation, but also on the number of cars and the number of streets.

The algorithm will terminate when the cars cover the total duration of the simulation.

Arguments

adj_list::AdjList{ModifiedStreet}: The adjacency list.

spreading_duration::Int: The duration for which cars will spread out, i.e. try to stick to streets directed close to an angle corresponding to their index. If not specified, it will be set to 1/10 of the total duration.

Returns

A Solution object.