HashCode2014Solution

Documentation for HashCode2014Solution.

HashCode2014Solution.AdjList

HashCode2014Solution.ModifiedJunction

HashCode2014Solution.create_adj_list

HashCode2014Solution.get_connectivity

HashCode2014Solution.get_direction

HashCode2014Solution.get_modified_junctions

HashCode2014Solution.get_modified_streets

HashCode2014Solution.solve_algorithm

HashCode2014Solution.AdjList — Type

struct AdjList{T}

A struct representing an adjacency list.

Fields

edges::Vector{Vector{T}}: The adjacency list.

total_duration::Int: The total duration of the city.

HashCode2014Solution.ModifiedJunction — Type

struct ModifiedJunction

A struct representing a junction. Similar to Junction from HashCode2014, but with an additional field index representing the index of the junction.

Fields

index::Int: The index of the junction.

latitude::Float64: The latitude of the junction.

longitude::Float64: The longitude of the junction.

HashCode2014Solution.create_adj_list — Method

create_adj_list(city::City)

Create an adjacency list based on the City object.

Arguments

city::City: The City object containing the streets.

Returns

An AdjList object.

HashCode2014Solution.get_connectivity — Method

get_connectivity(city::City, street::Street)

Returns the number of streets connected to the endpoint of a given street in a city.

Arguments

city::City: The city object.

street::Street: The street object.

Returns

Int: The number of streets connected to the endpoint of the given street.

HashCode2014Solution.get_direction — Method

get_direction(city::City, A::Int, B::Int)

Compute the angle of the street from junction A to junction B in a given city with respect to the north.

Arguments

city::City: The city object containing the junctions.

A::Int: The index of the starting junction.

B::Int: The index of the destination junction.

Returns

The direction in degrees from junction A to junction B.

HashCode2014Solution.get_modified_junctions — Method

get_modified_junctions(city::City)

Create a list of ModifiedJunction objects based on the City object.

Arguments

city::City: The City object containing the junctions.

Returns

An array of ModifiedJunction objects.

HashCode2014Solution.get_modified_streets — Method

get_modified_streets(city::City)

Create a list of ModifiedStreet objects based on the City object.

Arguments

city::City: The City object containing the streets.

Returns

An array of ModifiedStreet objects.

HashCode2014Solution.solve_algorithm — Method

solve_algorithm()

This function is used to solve the algorithm.